+++ src/gevent/thread.py
@@ -53,7 +53,10 @@ error = __thread__.error
 
 from gevent._compat import PYPY
 from gevent._util import copy_globals
-from gevent.hub import getcurrent, GreenletExit
+from gevent.hub import getcurrent
+from gevent.hub import GreenletExit
+from gevent.hub import sleep
+from gevent._hub_local import get_hub_if_exists
 from gevent.greenlet import Greenlet
 from gevent.lock import BoundedSemaphore
 from gevent.local import local as _local
@@ -89,23 +92,45 @@ class LockType(BoundedSemaphore):
 
     if PY3:
         _TIMEOUT_MAX = __thread__.TIMEOUT_MAX # python 2: pylint:disable=no-member
+    else:
+        _TIMEOUT_MAX = 9223372036.0
 
-        def acquire(self, blocking=True, timeout=-1):
-            # Transform the default -1 argument into the None that our
-            # semaphore implementation expects, and raise the same error
-            # the stdlib implementation does.
-            if timeout == -1:
-                timeout = None
-            if not blocking and timeout is not None:
-                raise ValueError("can't specify a timeout for a non-blocking call")
-            if timeout is not None:
-                if timeout < 0:
-                    # in C: if(timeout < 0 && timeout != -1)
-                    raise ValueError("timeout value must be strictly positive")
-                if timeout > self._TIMEOUT_MAX:
-                    raise OverflowError('timeout value is too large')
+    def acquire(self, blocking=True, timeout=-1):
+        # This is the Python 3 signature.
+        # On Python 2, Lock.acquire has the signature `Lock.acquire([wait])`
+        # where `wait` is a boolean that cannot be passed by name, only position.
+        # so we're fine to use the Python 3 signature.
 
-            return BoundedSemaphore.acquire(self, blocking, timeout)
+        # Transform the default -1 argument into the None that our
+        # semaphore implementation expects, and raise the same error
+        # the stdlib implementation does.
+        if timeout == -1:
+            timeout = None
+        if not blocking and timeout is not None:
+            raise ValueError("can't specify a timeout for a non-blocking call")
+        if timeout is not None:
+            if timeout < 0:
+                # in C: if(timeout < 0 && timeout != -1)
+                raise ValueError("timeout value must be strictly positive")
+            if timeout > self._TIMEOUT_MAX:
+                raise OverflowError('timeout value is too large')
+
+        acquired = BoundedSemaphore.acquire(self, blocking, timeout)
+        if not acquired and not blocking and getcurrent() is not get_hub_if_exists():
+            # Run other callbacks. This makes spin locks works.
+            # We can't do this if we're in the hub, which we could easily be:
+            # printing the repr of a thread checks its tstate_lock, and sometimes we
+            # print reprs in the hub.
+            # See https://github.com/gevent/gevent/issues/1464
+
+            # By using sleep() instead of self.wait(0), we don't force a trip
+            # around the event loop *unless* we've been running callbacks for
+            # longer than our switch interval.
+            sleep()
+        return acquired
+
+    # Should we implement _is_owned, at least for Python 2? See notes in
+    # monkey.py's patch_existing_locks.
 
 allocate_lock = LockType