+++ src/gevent/timeout.py
@@ -266,12 +266,16 @@ class Timeout(BaseException):
         # Internal use only in 1.1
         # Return an object with a 'cancel' method; if timeout is None,
         # this will be a shared instance object that does nothing. Otherwise,
-        # return an actual Timeout. Because negative values are hard to reason about,
+        # return an actual Timeout. A 0 value is allowed and creates a real Timeout.
+
+        # Because negative values are hard to reason about,
         # and are often used as sentinels in Python APIs, in the future it's likely
         # that a negative timeout will also return the shared instance.
-        # This saves the previously common idiom of 'timer = Timeout.start_new(t) if t is not None else None'
+        # This saves the previously common idiom of
+        # 'timer = Timeout.start_new(t) if t is not None else None'
         # followed by 'if timer is not None: timer.cancel()'.
         # That idiom was used to avoid any object allocations.
+
         # A staticmethod is slightly faster under CPython, compared to a classmethod;
         # under PyPy in synthetic benchmarks it makes no difference.
         if timeout is None: