+++ CHANGES.rst
@@ -29,6 +29,12 @@
   greenlets. Previously this was hidden by forcing buffering, which
   raised ``RuntimeError``.
 
+- Fix using monkey-patched ``threading.Lock`` and ``threading.RLock``
+  objects as spin locks by making them call ``sleep(0)`` if they
+  failed to acquire the lock in a non-blocking call. This lets other
+  callbacks run to release the lock, simulating preemptive threading.
+  Using spin locks is not recommended, but may have been done in code
+  written for threads, especially on Python 3. See :issue:`1464`.
 
 1.5a2 (2019-10-21)
 ==================